backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
chromadb==0.4.18
ollama==0.1.6
pymupdf==1.23.8
beautifulsoup4==4.12.2
langchain==0.1.0
langchain-community==0.0.10
python-dotenv==1.0.0
aiofiles==23.2.1
email-validator==2.1.0

